pyproject.toml
[tool.poetry]
name = "langgraph-tutorial"
version = "0.1.0"
description = ""
authors = ["Rajat Jain <155104750+rajat-hazlnut@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
assemblyai = "^0.33.0"
uvicorn = "^0.30.6"
uuid = "^1.30"
fastapi = "^0.112.2"
pyht = "^0.0.28"
langchain-core = "^0.2.38"
langchain-community = "^0.2.16"
langgraph = "0.2.14"
langchain-openai = "^0.1.23"
langchain = "^0.2.16"
twilio = "^9.2.4"
google-api-core = "^2.19.2"
google-api-python-client = "^2.143.0"
google-auth = "^2.34.0"
google-auth-httplib2 = "^0.2.0"
pytz = "^2024.1"
python-multipart = "^0.0.9"
ipython = "^8.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
